Added redhill core (forgot before) and fixed int bug in migrations

+++ db/migrate/20080625124657_create_phone_numbers.rb
@@ -1,9 +1,9 @@
 class CreatePhoneNumbers < ActiveRecord::Migration
   def self.up
     create_table :phone_numbers do |t|
-      t.int :phone_number_relationship_id
-      t.bigint :number
-      t.int :country_id
+      t.integer :phone_number_relationship_id
+      t.integer :number
+      t.integer :country_id
 
       t.timestamps
     end
